+++ services/gateway_manager/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 server.port=9015
 
 #####Microservice url configuration######
-microservices.zeta.service.url=http://localhost:8080
+microservices.zeta.service.url=http://10.213.43.90
 microservices.dpm.service.url=http://localhost:9010
 
 #####Ignite configuration######
